Define locale for format of expires string
A customer reported that on some devices the `expires` string is in the format 2021/07/19 08:03:07AM+00:00. The AM part not accepted by Transloadit's API. The customer also said that setting the device's time format to 24hrs fixed this problem. This patch attempts to fix this by setting a concrete locale.

This patch has not been tests as I currently do not have access to a testing device or simulator.

Ref: https://app.intercom.com/a/apps/qiqpfgjg/inbox/inbox/230775/conversations/26852700003272